+++ lib/src/aliasing.rs
@@ -2,29 +2,21 @@ use std::collections::HashMap;
 use std::fs::File;
 use std::io::prelude::*;
 
-use log::{info, warn};
+use log::info;
 
-use crate::{error::Error, settings::configuration_directories};
+use crate::error::Error;
+use crate::settings::Settings;
 
 /// Return the contents of the alias file, if it exists and can be parsed. \
 /// The file has to be located in `pueue_directory` and named `pueue_aliases.yml`.
-pub fn get_aliases() -> Result<HashMap<String, String>, Error> {
+pub fn get_aliases(settings: &Settings) -> Result<HashMap<String, String>, Error> {
     // Go through all config directories and check for a alias file.
-    let mut alias_file_path = None;
-    for directory in configuration_directories() {
-        let path = directory.join("pueue_aliases.yml");
-        if path.exists() {
-            alias_file_path = Some(path);
-        }
-    }
+    let path = settings.shared.alias_file();
 
     // Return early if we cannot find the file
-    let path = match alias_file_path {
-        None => {
-            info!("Didn't find pueue alias file.");
-            return Ok(HashMap::new());
-        }
-        Some(alias_file_path) => alias_file_path,
+    if !path.exists() {
+        info!("Didn't find pueue alias file at {path:?}.");
+        return Ok(HashMap::new());
     };
 
     // Read the file content
@@ -42,24 +34,23 @@ pub fn get_aliases() -> Result<HashMap<String, String>, Error> {
 
 /// Check if there exists an alias for a given command.
 /// Only the first word will be replaced.
-pub fn insert_alias(command: String) -> String {
+pub fn insert_alias(settings: &Settings, command: String) -> String {
+    // Get the first word of the command.
     let first = match command.split_whitespace().next() {
         Some(first) => first,
         None => return command,
     };
 
-    let aliases = match get_aliases() {
+    let aliases = match get_aliases(settings) {
         Err(err) => {
-            warn!("Failed to open aliases file: {err}");
+            info!("Couldn't read aliases file: {err}");
             return command;
         }
         Ok(aliases) => aliases,
     };
 
-    for (original, alias) in aliases.iter() {
-        if original == first {
-            return command.replacen(original, alias, 1);
-        }
+    if let Some(alias) = aliases.get(first) {
+        return command.replacen(first, alias, 1);
     }
 
     command